Completed implementation of messagelayer.
